feat(nullsearchservice): fix PR
