+++ src/Services/NullSearchService.php
@@ -1,37 +1,54 @@
 <?php
 
-namespace Algolia\SearchBundle\Service;
+namespace Algolia\SearchBundle\Services;
 
 use Algolia\AlgoliaSearch\RequestOptions\RequestOptions;
+use Algolia\AlgoliaSearch\Response\NullResponse;
 use Doctrine\Common\Persistence\ObjectManager;
 
-interface SearchServiceInterface
+/**
+ * This class aims to be used in dev or testing environments.
+ * It may be subject to breaking changes.
+ */
+final class NullSearchService implements SearchServiceInterface
 {
     /**
      * @param string $className
      *
      * @return bool
      */
-    public function isSearchable($className);
+    public function isSearchable($className)
+    {
+        return false;
+    }
 
     /**
      * @return array<int, string>
      */
-    public function getSearchables();
+    public function getSearchables()
+    {
+        return [];
+    }
 
     /**
      * @return array<string, array|int|string>
      */
-    public function getConfiguration();
+    public function getConfiguration()
+    {
+        return [
+            'batchSize' => 200,
+        ];
+    }
 
     /**
-     * Get the index name for the given `$className`.
-     *
      * @param string $className
      *
      * @return string
      */
-    public function searchableAs($className);
+    public function searchableAs($className)
+    {
+        return $className;
+    }
 
     /**
      * @param object|array<int, object>                      $searchables
@@ -39,7 +56,10 @@ interface SearchServiceInterface
      *
      * @return \Algolia\AlgoliaSearch\Response\AbstractResponse
      */
-    public function index(ObjectManager $objectManager, $searchables, $requestOptions = []);
+    public function index(ObjectManager $objectManager, $searchables, $requestOptions = [])
+    {
+        return new NullResponse();
+    }
 
     /**
      * @param object|array<int, object>                      $searchables
@@ -47,7 +67,10 @@ interface SearchServiceInterface
      *
      * @return \Algolia\AlgoliaSearch\Response\AbstractResponse
      */
-    public function remove(ObjectManager $objectManager, $searchables, $requestOptions = []);
+    public function remove(ObjectManager $objectManager, $searchables, $requestOptions = [])
+    {
+        return new NullResponse();
+    }
 
     /**
      * @param string                                         $className
@@ -55,7 +78,10 @@ interface SearchServiceInterface
      *
      * @return \Algolia\AlgoliaSearch\Response\AbstractResponse
      */
-    public function clear($className, $requestOptions = []);
+    public function clear($className, $requestOptions = [])
+    {
+        return new NullResponse();
+    }
 
     /**
      * @param string                                         $className
@@ -63,7 +89,10 @@ interface SearchServiceInterface
      *
      * @return \Algolia\AlgoliaSearch\Response\AbstractResponse
      */
-    public function delete($className, $requestOptions = []);
+    public function delete($className, $requestOptions = [])
+    {
+        return new NullResponse();
+    }
 
     /**
      * @param string                                         $className
@@ -74,7 +103,12 @@ interface SearchServiceInterface
      *
      * @throws \Algolia\AlgoliaSearch\Exceptions\AlgoliaException
      */
-    public function search(ObjectManager $objectManager, $className, $query = '', $requestOptions = []);
+    public function search(ObjectManager $objectManager, $className, $query = '', $requestOptions = [])
+    {
+        return [
+            new \stdClass(),
+        ];
+    }
 
     /**
      * @param string                                         $className
@@ -85,7 +119,12 @@ interface SearchServiceInterface
      *
      * @throws \Algolia\AlgoliaSearch\Exceptions\AlgoliaException
      */
-    public function rawSearch($className, $query = '', $requestOptions = []);
+    public function rawSearch($className, $query = '', $requestOptions = [])
+    {
+        return [
+            'result' => [],
+        ];
+    }
 
     /**
      * @param string                                         $className
@@ -96,5 +135,18 @@ interface SearchServiceInterface
      *
      * @throws \Algolia\AlgoliaSearch\Exceptions\AlgoliaException
      */
-    public function count($className, $query = '', $requestOptions = []);
+    public function count($className, $query = '', $requestOptions = [])
+    {
+        return 0;
+    }
+
+    /**
+     * @param object $entity
+     *
+     * @return bool
+     */
+    public function shouldBeIndexed($entity)
+    {
+        return false;
+    }
 }
